responsive

+++ src/components/RemoveModal.js
@@ -1,22 +1,29 @@
 import { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux'
-import { removeCurrency } from '../store/actions'
+import { removeCurrency, flushCurrency } from '../store/actions'
 import { removeModal as styleRemove, buttons as styleButtons } from '../css/modals.module.css'
+import { removeAllBtn as removeAllStylesBtn } from '../css/user-panel.module.css'
 import { burntBtn } from '../css/main.module.css'
 
-function RemoveModal({ currency }) {
+function RemoveModal({ currency, removeAll = false }) {
     const dispatch = useDispatch()
     const authUser = useSelector(store => store.auth)
     const [isVisible, setIsVisible] = useState(false)
     const removeCurrencyHandler = (select) => {
-        if(select === 'yes') {
-            dispatch(removeCurrency({user: authUser.name, code: currency.code}))
+        if(removeAll && select === 'yes') {
+            dispatch(flushCurrency(authUser.name))
+        } else {
+            if(select === 'yes') {
+                dispatch(removeCurrency({user: authUser.name, code: currency.code}))
+            }
         }
         setIsVisible(false)
     }
     return(
         <>
-        <button className={burntBtn} onClick={() => setIsVisible(true)}>remove</button>
+        <button className={removeAll ? removeAllStylesBtn : burntBtn} onClick={() => setIsVisible(true)}>
+            {removeAll ? 'remove all' : 'remove'}
+        </button>
         {isVisible ?
             <div className={styleRemove}>
                 <div>
